requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.1.1
Django==3.1.1
django-cors-headers==3.5.0
django-rest-knox==4.1.0
djangorestframework==3.12.1
filter==0.0.0.20200724
httpie==2.2.0
idna==2.10
Markdown==3.2.2
pycparser==2.20
Pygments==2.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
